requirements.txt
requests>=2.25.0
click>=8.0.0
colorama>=0.4.4
tabulate>=0.8.9
prompt_toolkit>=3.0.20
rich>=10.0.0
openai>=1.60.0
google-api-python-client>=2.0.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=0.4.6
